test(cli-run): split event handler specs by concern
